requirements.txt
langchain-community
openai
python-dotenv